pyproject.toml
[project]
name = "mcenter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "asgiref==3.11.0",
    "django==6.0.1",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "pillow==12.1.0",
    "pyjwt==2.11.0",
    "python-dotenv==1.2.1",
    "ruff>=0.15.1",
    "sqlparse==0.5.5",
    "tzdata==2025.3",
]
[tool.ruff.lint.isort]
known-first-party = ["appointment", "users", "mcenter"]